refactor: util 파일 util 폴더로 이동
