+++ src/utils/codeValidateUtils.ts
